refactor(product-projection-alternative): use the new specialized build signature
